+++ app/build.gradle.kts
@@ -160,7 +160,6 @@ dependencies {
     implementation(projects.features.languages)
     implementation(projects.features.media)
     implementation(projects.features.navigationSuite)
-    implementation(projects.features.pdf)
     implementation(projects.features.resource)
     implementation(projects.features.settings)
     implementation(projects.features.share)
@@ -172,6 +171,7 @@ dependencies {
     implementation(projects.services.lessons.impl)
     implementation(projects.services.media.impl)
     implementation(projects.services.media.ui)
+    implementation(projects.services.pdf.impl)
     implementation(projects.services.prefs.impl)
     implementation(projects.services.resources.impl)
     implementation(projects.services.storage.impl)
