+++ services/pdf/impl/src/main/AndroidManifest.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+    <application>
+        <meta-data
+            android:name="pspdfkit_license_key"
+            android:value="${psPdfKitKey}" />
+    </application>
+</manifest>
