+++ services/pdf/impl/src/main/kotlin/ss/services/pdf/impl/PdfReaderImpl.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025. Adventech <info@adventech.io>
+ * Copyright (c) 2026. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package app.ss.pdf
+package ss.services.pdf.impl
 
 import android.content.Context
 import android.net.Uri
