+++ settings.gradle.kts
@@ -57,7 +57,6 @@ include(
     ":features:feed",
     ":features:media",
     ":features:navigation-suite",
-    ":features:pdf",
     ":features:resource",
     ":features:settings",
     ":features:share",
@@ -91,6 +90,7 @@ include(
     ":services:lessons:impl",
     ":services:media:impl",
     ":services:media:ui",
+    ":services:pdf:impl",
     ":services:prefs:impl",
     ":services:resources:api",
     ":services:resources:impl",
